# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveRequest": {
        "description": "POST body for approve: promotes the staging row to ``reports``.\n\n``decision_reason`` is intentionally not declared here. Clients that\nsend it get it silently dropped (plan §2.1 D1 \"approve 시 무시\"):\napprove has no semantic \"reason\" — the approval itself is the\naction. ``notes`` is optional and lands in\n``audit_log.diff_jsonb.reviewer_notes``.",
        "properties": {
          "decision": {
            "const": "approve",
            "title": "Decision",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ApproveRequest",
        "type": "object"
      },
...
      "RejectRequest": {
        "description": "POST body for reject: soft-rejects the staging row.\n\n``decision_reason`` is required and non-empty after whitespace\nstripping — a reject with no reason is unsafe (audit forensics\nrequire the cause). ``notes`` is optional and audit-only.\n\nReopen (``rejected → pending``) is a future admin-only action, out\nof PR #10 scope (plan §3 Out of scope, §2.2 C).",
        "properties": {
          "decision": {
            "const": "reject",
            "title": "Decision",
            "type": "string"
          },
          "decision_reason": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Decision Reason",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "decision",
          "decision_reason"
        ],
        "title": "RejectRequest",
        "type": "object"
      },
...
      "ReportListResponse": {
        "description": "Keyset-paginated response for ``GET /api/v1/reports``.\n\n``next_cursor`` is ``None`` on the final page. The cursor is\nopaque — see ``api.read.pagination``.",
        "examples": [
          {
            "items": [
              {
                "id": 42,
                "lang": "en",
                "published": "2026-03-15",
                "source_id": 7,
                "source_name": "Mandiant",
                "title": "Lazarus targets South Korean crypto exchanges",
                "tlp": "WHITE",
                "url": "https://mandiant.com/blog/lazarus-2026q1",
                "url_canonical": "https://mandiant.com/blog/lazarus-2026q1"
              }
            ],
            "next_cursor": "MjAyNi0wMy0xNXw0Mg"
          },
          {
            "items": []
          }
        ],
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ReportListResponse",
        "type": "object"
      },
      "ReviewDecisionResponse": {
        "description": "200 body for both approve and reject. Minimal shape by design\n(plan §8 #2 lock): clients reading from ``staging`` / ``reports``\ndirectly don't need the service to re-serialize the full row.\n\n``report_id`` is populated on approve (new or attached_existing),\n``None`` on reject.",
        "properties": {
          "report_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "staging_id": {
            "title": "Staging Id",
            "type": "integer"
          },
          "status": {
            "enum": [
              "promoted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "staging_id",
          "report_id",
          "status"
        ],
        "title": "ReviewDecisionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Implementation-prep OpenAPI skeleton for the DPRK CTI platform.",
    "title": "DPRK CTI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reports": {
      "get": {
        "description": "Returns reports sorted by `published DESC, id DESC` (plan D11). Pagination is keyset-style — clients pass the opaque `next_cursor` from the previous page to continue (plan D3). Filter params per plan D5 with OR semantics inside a repeatable (`tag=a&tag=b` matches either) and AND semantics across different filters (`tag=a&source=Mandiant` requires both). JOIN uses EXISTS so a report with two matching tags still surfaces as one row.",
        "operationId": "list_reports_endpoint_api_v1_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "happy": {
                    "summary": "First page, next page exists",
                    "value": {
                      "items": [
                        {
                          "id": 42,
                          "lang": "en",
                          "published": "2026-03-15",
                          "source_id": 7,
                          "source_name": "Mandiant",
                          "title": "Lazarus targets South Korean crypto exchanges",
                          "tlp": "WHITE",
                          "url": "https://mandiant.com/blog/lazarus-2026q1",
                          "url_canonical": "https://mandiant.com/blog/lazarus-2026q1"
                        }
                      ],
                      "next_cursor": "MjAyNi0wMy0xNXw0Mg"
                    }
                  },
                  "last_page": {
                    "summary": "Final page — no next_cursor",
                    "value": {
                      "items": []
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ReportListResponse"
                }
              }
            },
            "description": "One page of reports with next_cursor metadata."
          },
          "401": {
            "description": "Missing or invalid session cookie"
          },
          "403": {
            "description": "Role not analyst / researcher / policy / soc / admin"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "query",
                        "cursor"
                      ],
                      "msg": "malformed cursor",
                      "type": "value_error.malformed_cursor"
                    }
                  ]
                }
              }
            },
            "description": "Invalid query param — malformed cursor, empty filter value, invalid ISO date, out-of-range limit, etc. Plan D12 locks uniform 422 with no silent ignore."
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": "rate_limit_exceeded",
                  "message": "60 per 1 minute"
                }
              }
            },
            "description": "Rate limit exceeded — 60/min/user read bucket (plan D2). Per-route bucket: draining this does NOT consume the /reports POST-review 30/min mutation bucket."
          }
        },
        "summary": "List reports (keyset-paginated, filterable)",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/review/{staging_id}": {
      "post": {
        "description": "§5.3 / §9.3 Approve or reject a staging row.\n\nCaller owns the transaction: the handler wraps the entire\ndecision (service call + any side effects) in a single\n``async with session.begin():``. Plan §2.2 A locks this boundary\nto the handler so the rollback path on service exceptions is\nuniform regardless of which error surfaces.\n\nException → HTTP mapping (plan §2.2 B response enum lock +\nGroup D service-level split):\n- ``StagingNotFoundError``           → 404\n- ``StagingAlreadyDecidedError``     → 409 with ``AlreadyDecidedError``\n                                     body (current_status narrowed\n                                     to promoted|rejected)\n- ``StagingInvalidStateError``       → 422 (current_status includes\n                                     approved|error — not exposed\n                                     via the 409 DTO narrowing)\n- ``PromoteValidationError``         → 422 with reason surface\n\nThe error bodies return via ``JSONResponse`` directly (not via\n``HTTPException``) so we can match the ``AlreadyDecidedError`` DTO\nshape exactly, without FastAPI's default ``{\"detail\": ...}``\nwrapper obscuring the contract.",
        "operationId": "review_staging_api_v1_reports_review__staging_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "staging_id",
            "required": true,
            "schema": {
              "title": "Staging Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApproveRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RejectRequest"
                  }
                ],
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid session cookie"
          },
          "403": {
            "description": "Authenticated role is not analyst / researcher / admin"
          },
          "404": {
            "description": "Staging row not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlreadyDecidedError"
                }
              }
            },
            "description": "Row already promoted or rejected"
          },
          "422": {
            "description": "Validation error — staging row is in an unhandleable state (approved / error) or is missing NOT NULL fields"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": "rate_limit_exceeded",
                  "message": "30 per 1 minute"
                }
              }
            },
            "description": "Rate limit exceeded — 30/min/user (plan D2 mutation bucket)."
          }
        },
        "summary": "Review Staging",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "description": "Shallow-joined detail view of a single report. Plan D9 caps the `linked_incidents` collection at 10 newest-reported entries; reports with > 10 incident links return only the top 10 (ordered `reported DESC NULLS LAST, id DESC`). `tags`, `codenames`, and `techniques` are flat string lists. Returns 404 when the report id is unknown.",
        "operationId": "get_report_detail_endpoint_api_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            },
            "description": "Report detail with shallow joins (plan D9)."
          },
          "401": {
            "description": "Missing or invalid session cookie"
          },
          "403": {
            "description": "Role not analyst / researcher / policy / soc / admin"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "report not found"
                }
              }
            },
            "description": "Report id not found"
          },
          "422": {
            "description": "Invalid path parameter — non-integer report_id. Plan D12 uniform 422 (FastAPI path-param validation)."
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": "rate_limit_exceeded",
                  "message": "60 per 1 minute"
                }
              }
            },
            "description": "Rate limit exceeded — 60/min/user read bucket (plan D2). Per-route bucket: this endpoint's drain does NOT consume the /reports list bucket."
          }
        },
        "summary": "Get one report with related joins and capped linked incidents",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{report_id}/similar": {
      "get": {
        "description": "§5.2 pgvector similarity search for related reports.\n\nStub: returns 501 until Phase 3 slice 1 Group B (plan D2 + D8 + D10).",
        "operationId": "similar_reports_api_v1_reports__report_id__similar_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "default": 10,
              "title": "K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Similar Reports",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}